apfs: begin work on apfs module
Create a simple module capable only of mounting and unmounting an apfs
filesystem. Some data from the volume headers is printed for testing.
The plan is for this to grow into a fully functional read-only module.

The apfs specification is not public, and this work is based on reverse
engineering done by others [1]. There is still plenty of uncertainty
regarding the disk layout.

[1] Hansen, K.H., Toolan, F., Decoding the APFS file system, Digital
    Investigation (2017), http://dx.doi.org/10.1016/j.diin.2017.07.003

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>