apfs: populate vfs inode on lookup
So far the ->lookup() method produced only a shell of an inode, for the
purpose of testing the cnid lookup. Read the actual inode metadata from
disk and use it to initialize the vfs inode. With this change stat will
now work (almost) properly.

There is some remaining confusion about the on-disk inode structures.
Most importantly, many files do not seem to report their size at all.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>