apfs: reorganize b-tree and table lookups
The apfs_btom_read_table(), apfs_cat_resolve(), and apfs_cat_get_data()
functions are very similar to each other, so reorganize the code around
a query structure to be able to reuse some of it.

In the process make them a bit more solid by adding the often forgotten
length checks. And no longer demand the release of the returned table
from apfs_cat_get_data() in case of failure.

Depressingly enough, this patch does not actually decrease the line
count. But this is mostly because of excessive comments, and the query
code will be used more in the future.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>