apfs: prevent potential out-of-bounds table reads
An apfs table could be crafted so that it claims to have entries beyond
the limits of the block. Add boundary checks in apfs_table_locate_key(),
apfs_table_locate_data() and apfs_read_table().

For the protection to be effective, the callers of these functions must
only operate within the limits provided by them. Most were already doing
this correctly; fix the one exception in apfs_count_used_blocks().

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>