apfs: add support for hard links
This change was mostly trivial; as far as we are concerned, the only
thing that sets apart files with hard links is the length of their
dentries.

A remaining issue is that directories do not report their link count
(that is, the number of subdirectories plus two) among the inode data;
it seems that to provide it we would have to actually count the
subdirectories one by one. The HFS and HFS+ modules do not bother and
just set the link count to one, which should be good enough for now.

By the way, stop printing the link (or child) count every time
apfs_iget() is called. This was intended for testing but was never very
useful.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>