apfs: search dentry keys by hash instead of name
The paper this module was following [1] mistakenly claimed that catalog
keys were ordered by their name, when in fact they are ordered by a hash
of its normalized form. This information was obtained from the apfs-fuse
project [2].

The result was that some files would show up under ls but be declared
nonexistent by stat. I have been aware of this issue for a while, but
before we had support for hard links I thought that was the reason.

Any simple fix to this bug would be very ugly, so instead restructure
the query code once again. Use in-memory key structures for comparisons
instead of pointers to data on disk, and add helper functions for
initializing them. Also declare separate structures for the different
types of on-disk keys. Move all the related code to the new key.c and
key.h files.

[1] Hansen, K.H., Toolan, F., Decoding the APFS file system, Digital
    Investigation (2017), http://dx.doi.org/10.1016/j.diin.2017.07.003
[2] https://github.com/sgan81/apfs-fuse

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>