apfs: stop reading link count for special files
It seems that apfs does not allow hard links for special files, with the
possible exception of symlinks. The link count field in those inodes is
used for a different purpose, overflowing the i_nlink field of the vfs
inode and causing calls to stat to fail.

Just leave i_nlink at one, as we do for directories.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>