apfs: define EFSCORRUPTED error code
Follow xfs in returning a specific error code when filesystem corruption
is detected. Until now we were using EINVAL is such situations, which is
confusing and not very appropriate.

No return codes have been changed in the apfs_fill_super() function
because it is too long and messy; I would rather break it up before.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>