apfs: allow queries for best match
Back when the query code was first written, I was under the impression
that the named entries in a table were ordered by their name: they are
actually ordered by a hash. From this confusion I concluded that the
entries in a leaf node were stored out of order.

Now that I know better, create a flag to let queries know if they should
return only exact matches, or if they should return the last entry that
comes before the one requested. This would have been much harder to do
if the records were not in order, and it will be useful when dealing
with extents.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>