apfs: implement file read
Create an apfs_get_block() function to find and map the extent to which
a given file block belongs. Use this function to define a file's address
space operations, and thus provide the read functionality.

There is still much to figure out; most importantly, inodes that do not
report their length in the expected way will not be read at all.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>