apfs: make xattr search case-sensitive
After the order of the records in the leaf nodes was made relevant for
queries, some named attributes became unreadable. It seems that xattrs
are case-sensitive even if the filesystem is not, and the records are
ordered accordingly. So take this into consideration when comparing
their keys.

Also, now that I have a better idea how xattrs are ordered, remove a few
comments with speculation on the subject.

Fixes: dcf639c3f1bf ("apfs: allow queries for best match")
Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>