apfs: parse xattr data correctly
Until now calls to getxattr were simply returning the catalog data for
the named attribute. This is not correct, since the data always has a
header before the value, and in some cases the value is kept on extents.

Create new structures for the xattr data in both cases, and use them to
read the actual values.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>