apfs: fix hashing of utf-8 filenames
We have no real support for unicode filenames yet, but handling them
is very simple when they are stored in decomposed form (and not
capitalized). The only thing standing in the way is the hashing done
by apfs_init_key(), which expects single byte characters. So fix that.

This change requires passing the name length to the function; the number
of hardly ever used parameters it has is becoming a bit absurd. I should
probably give up on trying to initialize all key types with the same
function.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>