apfs: fix length of mapping from apfs_get_block()
If the length of an extent is not a multiple of the block size, round
it up for the buffer head. Also, never map more blocks than requested
by the vfs layer. This has not caused any problems so far, but it seems
reasonable and most in-tree filesystems do it.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>