apfs: parse inode data correctly
Before this patch I did not understand why some inodes seemed to have
more fields than others, and how those fields were organized. Some of
them held important information about the file size, so I put a hacky
solution in place to read them whenever they seemed to fit.

Now I found an actual explanation in apfs-fuse [1], so clean up the code
and get it to work properly in all cases. Note that the size is the only
one of these optional inode attributes that we bother to read; I do not
yet know what the others are for.

[1] https://github.com/sgan81/apfs-fuse

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>