apfs: support unicode filenames
Add support for case folding and nfd decomposition of filename strings
according to Unicode 9.0. A script was used to generate and test the
data arrays [1]. The intention was to follow the hfsplus implementation,
but its many bugs made that impossible.

This patch causes a significant performance hit, which can be prevented
by skipping the trie searches in the common ascii case. The next one
will take care of this.

[1] https://github.com/eafer/apfs-ucd-parser

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>