apfs: Add container superblock checksum verification
Based on the Fletcher64 algorithm from drivers/nvdimm.  Nevertheless,
differently from nvdimm, which follows a variant of Fletcher64 where the
Fletcher's modulo M is 2^K, apfs uses M=(2^k)-1. This is why we can't
use the overflow wrap of the counter to do the modulo like nvdimm, and
need to do it explicitly with do_div().

On the other hand, we guarantee that sum1 and sum2 will not be
overflowed during the summation, because blocksize is limited to 2^16,
so we don't need to worry about wrapping during the loop.

Signed-off-by: Gabriel Krisman Bertazi <krisman@collabora.co.uk>
[ernesto.mnd.fernandez@gmail.com: changed style of comment]
Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>