apfs: Protect apfs_table with kref
Instead of knowing beforehand which tables can or cannot be freed on
each situation, protect them all with reference counting.

In the current code, we guarantee that the tables are not going to be
freed while accessed by different threads, so it is safe to avoid
locks in the table structure as described in Documentation/kref.txt,
rule 3 for now.

Signed-off-by: Gabriel Krisman Bertazi <krisman@collabora.co.uk>
[ernesto.mnd.fernandez@gmail.com: obsolete comments removed]
Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>