apfs: Avoid tail recursion on btree lookup
Recursion in kernel space is usually undesirable, so let's avoid it when
traversing trees.

There is still one recursive call when translating child virtual ids to
physical ids, but it is assured to be a single-level recursion by this
patch, because this lookup is always type APFS_QUERY_OMAP, which is
implemented by the non-recursive path.

Signed-off-by: Gabriel Krisman Bertazi <krisman@collabora.co.uk>
Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>