apfs: don't round up the length of xattr extents
Badly implemented rounding causes apfs_xattr_extents_read() to expect
an extra block at the end of every extent.  This will be a problem for
xattrs with more than one.

The extent length must always be a whole number of blocks in a sane
filesystem, so just work under that assumption and don't attempt any
rounding.

Fixes: 79c1c25a3579 ("apfs: parse xattr data correctly")
Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>
Reviewed-by: Gabriel Krisman Bertazi <krisman@collabora.co.uk>