apfs: optimize file reading performance
The apfs_get_block() function always returned a length of one block,
even when asked to map multiple blocks: this is due to the map_bh()
function, which replaces the b_size field of a buffer_head structure
with the filesystem block's size, even if the data spans over multiple
blocks.

This led to reading the whole file one block at a time, and thus calling
apfs_btree_query() for each block, while we could read multiple blocks
on each iteration, and have a single query for all these blocks.

This patch makes sure apfs_get_block() returns the actual length it was
supposed to map whenever possible, so that we avoid doing too much
queries.

Signed-off-by: Arnaud Ferraris <arnaud.ferraris@collabora.com>
Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>