apfs: fix length check for xattr values
According to the official APFS reference, the xdata_len field of a xattr
record has no meaning for xattrs with a data stream.  Some length checks
were using that field, so fix them.

In practice, the Apple implementation was actually setting xdata_len to
the expected value, so this patch should make no difference.

Reported-by: Arnaud Ferraris <arnaud.ferraris@collabora.com>
Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>
Reviewed-by: Arnaud Ferraris <arnaud.ferraris@collabora.com>