apfs: fix size as reported by 'du'
The 'du' command (as well as 'ls' when used on a directory) expects the
blocks count reported by stat() to be based on a block size of 512
bytes, no matter what is the actual file system block size.
The i_blocks field of an inode should therefore contain the number of
512-bytes blocks the file would use, instead of the number of actual
file system blocks.

This patch allows 'du' and 'ls' to report correct sizes.

Signed-off-by: Arnaud Ferraris <arnaud.ferraris@collabora.com>
Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>