apfs: add support for cloned files
When copying a file, APFS doesn't copy its data, but rather creates a
new inode using the same extents as the original inode. Looking for an
extent with the ID of the new inode will give an invalid result.

In order to properly access cloned files, the file extents should
therefore be queried based on the private_id rather than the inode
number.

Signed-off-by: Arnaud Ferraris <arnaud.ferraris@collabora.com>
[ernesto.mnd.fernandez@gmail.com: 'extent_id' renamed to 'i_extent_id']
Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>