apfs: change i_extent_lock from mutex to spinlock
Following HFS+, a mutex is currently used to protect the extent cache.
This makes sense in a filesystem that supports writes, because reading a
new record into the extent cache may require sleeping while the old one
is committed to disk.

But the apfs module will remain read-only for the foreseeable future,
and it's better to write code that adjusts to that.  So, use a spinlock.

Suggested-by: Gabriel Krisman Bertazi <krisman@collabora.com>
Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>
Reviewed-by: Gabriel Krisman Bertazi <krisman@collabora.co.uk>