apfs: rework the apfs_key structure
The apfs_key structure is used only for comparisons during a query, so
the meaning of most of the fields doesn't matter, just the value.  The
'offset' and 'hash' fields are never set at the same time (and have the
same significance), so merge them into one.  Since it lacks any real
meaning, just call the new field 'number'.

Get rid of the messy apfs_init_key() and replace it with functions
specific to each record type.  Also rework how multiple queries work, by
rewriting the fields to ignore while reading a key from disk.  With this
change we can simplify the query that reads all extents of a xattr.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>