apfs: support 64-bit inode numbers on 32-bit archs
On a 32-bit architecture, the 64-bit inode numbers of apfs may overflow
the i_ino field of the vfs inode.  Currently, the module would respond
with an odd EOVERFLOW error.  This is unlikely to ever happen, but since
the inode numbers don't seem to be reused I'd rather deal with it, just
in case.

So add full support by means of iget5_locked() and an extra i_ino field
on apfs_inode_info, for 32-bit architectures only.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>