apfs: simplify bounds check for volume number
Before doing anything else, apfs_map_volume_super() must check that the
volume number specified by the user is not so large as to force an
out-of-bounds read.  The specification only allows 100 volumes, which
will always fit in a single block.  So, just check that the volume
number is below 100.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>