apfs: verify checksum of all metadata
Verify the checksums for not only the container superblock, but also the
volume superblock and the object maps.

As for the b-tree nodes, checking them has a big impact in performance,
probably because too many are being read repeatedly.  This will
hopefully improve in the future, but for now provide a mount option to
let the user decide what to do.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>