apfs: ban empty b-tree nodes
A multiple query that reaches an empty node simply assumes that all of
its records were already read, and continues with the next one.  A
crafted filesystem with a large number of empty nodes could keep this
spinning for a while.

I don't think there is any legitimate reason for empty nodes to exist on
disk, so just declare them corrupted in apfs_node_is_valid().

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>