apfs: begin work on write support
Start a new transaction on mount, copying the whole descriptor and data
areas.  Update the xids, oids and checksums of every object, as well as
the address reported by each checkpoint mapping.  Commit everything on
unmount by writing the new checkpoint superblock, with the correct
checksum.

The plan is to make all writes happen as part of this one transaction.
Of course this means that all changes will be lost in case of a crash,
but better keep things simple for now.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>