apfs: implement block allocation
Define the on-disk structures for the space manager in a new spaceman.h
header file; complete apfs_read_spaceman() and move it into spaceman.c.
With this in place, implement apfs_spaceman_allocate_block(), which
allocates a single block, as well as any more that are needed for the
spaceman metadata itself.

To test the new function, allocate a new block for the container's
object map.

I would have preferred to implement just the block allocation in this
patch, which is messy enough, but it's necessary to release the old
blocks for the filesystem to remain consistent.  Don't mark them as free
in the bitmaps yet, just add them to the free queue b-tree using the new
apfs_free_queue_insert().  To keep things simple, don't implement a
full-featured apfs_btree_insert() yet; this first draft will hit ENOSPC
very soon, but it's good enough for testing.

Free queue queries are needed here, so implement the key functions
apfs_read_free_queue_key() and apfs_init_free_queue_key(); also adjust
apfs_node_locate_data(), and allow a return value of 0 to be valid, as
expected for ghost records.

To know where to put the new entry, apfs_btree_insert() expects the
index reported by apfs_node_query() to always come right before the
correct location; set it to -1 on ENODATA instead of ignoring it.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>