apfs: allow multiple transactions per mount cycle
Instead of having a single transaction from mount to unmount, allow
transactions to be started and committed at will.  Also provide a means
to abort an ongoing transaction, which makes cleanup on failure easier
and more reliable.

Keep a list of transaction buffers with apfs_transaction_join(), and
flag the buffers that need a checksum instead of setting it right away.
This is now done during the commit, which should save time if the block
is modified more than once before.

Put all the new transaction code in a transaction.c file, and move the
checkpoint functions in there as well, since a new checkpoint is needed
for every new transaction.

Despite the new capabilities, the module still performs only one
transaction in the mount cycle (to update the volume unmount time), so
ignore the issue of locking for now; we'll get to it soon.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>