apfs: implement a simple locking scheme
For now we write only one transaction per mount cycle, but that should
be changing soon.  Since efficiency is not a priority, use a single
read-write semaphore to prevent all races; lock it for writing every
time a transaction is started, and unlock it when it ends.

Take the lock for reading at all points where it's necessary.  Since
apfs_xattr_get() is also called from inside the module, define
__apfs_xattr_get() to perform the same task without locking.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>