apfs: implement ->mknod() method
Implement the ->mknod() directory operation that creates special files.
Perform all writes immediately, putting no in-memory inodes in a dirty
state.  No cleanup of the parent inode is performed if the transaction
commit fails; btrfs seems to force the filesystem read-only in this
situation, so that should be considered in the future.

This patch requires adjustments to several existing functions.

Rework apfs_btree_insert() to allow keys and values of variable size,
such as those from the catalog tree.  While we are at it, also add
support for non-ghost fixed size records, which are not needed yet.

Extend the transaction code to remember the old root nodes for the
volume's catalog and omap trees, and to restore them in case of failure.
Reorder apfs_put_super() so that the nodes are not put until the
transaction is done.

To gain write access to these roots, add a 'write' parameter to
apfs_read_omap() and apfs_read_catalog(), in a similar style to
apfs_read_object_block(), for example.

Finally, set the superblock time granularity to nanoseconds on
apfs_fill_super(), so that current_time() will work as expected.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>