apfs: implement ->mkdir() directory operation
Following the hfsplus module, reuse apfs_mknod() to also implement the
->mkdir() directory operation.

Three small adjustments are required for this to work.  Move the setting
of the new inode's operations to apfs_new_inode(), so that all inode
types can be handled at once with apfs_inode_set_ops().  Don't set the
on-disk nlink field in apfs_build_inode_val() for directory inodes; they
report their child count instead.  And update the directory count in the
volume superblock, as appropriate.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>