apfs: implement ->link() directory operation
A hard link in APFS requires more than just a dentry record: sibling
links must be created to allow listing all links for an inode, and also
sibling maps, which serve no clear purpose.  Add support for those
record types.  Each dentry also needs an extended field that reports
its sibling id, so rework apfs_create_dentry_rec() to deal with that.

Files with a single link don't need sibling records, but change
apfs_mknod() to create them anyway.  With this adjustment ->link() is
much easier to get to work, but won't always be compatible with other
implementations of the filesystem, so it will need to be improved in the
future.

With that in place, implement the ->link() directory operation.  Besides
the calls to apfs_create_sibling_recs() and apfs_create_dentry_rec(),
a bit of ugly code is needed to update the link count in the inode
record.  I think it might be best to later move this stuff into a
function akin to ext4_mark_inode_dirty().

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>