apfs: use a wrapper to get the id of an inode
Write a simple apfs_ino() inline function that returns the 64-bit id of
an inode, hiding away the special needs of 32-bit architectures in this
matter.  Use it to get rid of the repetitive preprocessor macros in
dir.c and inode.c.

This issue had been neglected entirely in apfs_inode_lookup(), so take
this opportunity to get it fixed.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>