apfs: don't read the inode in apfs_inode_lookup()
After a successful search, make apfs_inode_lookup() return the query
structure instead of setting the inode fields; let the caller take care
of that.  This will allow us to reuse the lookup code for writes as
well.

To make this change explicit, declare the inode parameter as a constant.
As needed, propagate this change to the apfs_ino() helper.

Place the corruption alert inside apfs_inode_from_query(), since it's
the function that actually runs the checks.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>