apfs: move inode record updates to new function
Define apfs_update_inode() to take care of updating the on-disk record
for a modified in-memory inode.  For now focus only on the fields needed
by apfs_link() and apfs_create_dentry_rec(); simplify their code with
this new function.

In apfs_link(), also replace the calls to inode_inc_link_count() and
inode_dec_link_count() with inc_nlink() and dec_nlink(), respectively.
This should have been the case from the beginning: there is no need to
dirty the inode if the metadata is written to disk right away.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>