apfs: implement ->write_inode() method
Use apfs_update_inode() to implement the ->write_inode() superblock
operation.  Don't add support for all the inode fields yet, we'll deal
with that soon.

Note that ext4 seems to believe that this method might be called with an
open handle on a transaction.  That would cause a deadlock, but ignore
it for now.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>