apfs: don't give sibling records to a single link
A file with a single link does not require a sibling id, or sibling
records; the module makes them to avoid having to modify the records for
the primary link when a second one is added.  This is wasteful, and it's
also insufficient to deal with primary links created by other versions
of APFS.

So start doing this correctly.  To modify the dentry record, just delete
it using a first draft of apfs_btree_remove(), and then make it anew.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>