apfs: implement ->unlink() directory operation
Implement the ->unlink() directory operation.  It mainly just deletes
the dentry and sibling records, but two complications exist.

Inode records in APFS hold information about the primary link, so when
this link is deleted the inode needs to be modified as well.  Extend
apfs_update_inode() to support this use case.  Since the length of the
name is not fixed, the modified inode record may not fit in the same
location; for now just remove it and insert it again in all cases.  This
requires small tweaks to apfs_btree_remove() and apfs_btree_insert(), so
that queries can pass through them and still point to the right spot.

The other issue is that inodes with no links are not being deleted here,
which leaves them orphaned.  ->evict_inode() is not implemented yet, but
even if it was there could still be a crash before the file is closed.
The official reference says nothing on this matter, but we are allowed
to put anything we want in private-dir, so relink the orphaned inodes in
there for now.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>