apfs: clean up mess for 64-bit inode numbers
Having separate code to deal with inode numbers in 32-bit architectures
has proved unwise; there is no verified performance benefit from this
and now, after several commits with no testing, the 32-bit code won't
even build.  Fix this with a new const qualifier in the argument to
APFS_I(), and from now on just use the 32-bit version of everything.

It's important to always read the inode number with apfs_ino(), to avoid
getting confused and working with the potentially overflowing i_ino;
remove a few remaining instances of this bug.  Also start setting the
number by means of apfs_set_inode(), but don't touch the hacky code for
private-dir; I intend to turn it into a real inode in the future.

Finally, deal with the TODO regarding insert_inode_locked(): define
apfs_insert_inode_locked() as a wrapper to insert_inode_locked4(),
implementing the same functionality as insert_inode_locked() but with
full support for 32-bit architectures.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>