apfs: set filesystem read-only on commit failure
Callers don't properly reset in-memory structures after a transaction
commit fails.  For example, apfs_mknod() neglects to reduce the child
count of the parent directory.  This could be fixed of course, but it
seems wasteful: transaction commits can't hit ENOSPC and are unlikely
to hit ENOMEM, so they will typically not fail.

BTRFS always sets the filesystem read-only after aborting a transaction.
We can't follow that example in general because we don't yet preallocate
space, but there is no reason not to do it in this situation.  Adjust
apfs_transaction_start() and apfs_transaction_abort() to deal with these
unexpected read-only filesystems.

Another benefit of this change is that we no longer need to assume that
the checkpoint superblock has synced correctly; check for errors and
fail the commit as usual.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>