apfs: don't reset nlink on apfs_mknod() cleanup
If a fault is injected to apfs_transaction_commit(), creating a file
will deadlock.  The reason is that discard_new_inode() will call iput(),
which involves ->evict_inode() and will try to start a new transaction
that deletes all records for the supposed orphan.  This makes no sense;
avoid resetting the inode link count, and let apfs_transaction_abort()
handle the on-disk cleanup.

It should be noted that all other filesystems I checked do clear nlink
before discard_new_inode(), so there might be an important reason for
this that I'm not seeing.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>