apfs: use dir_emit_dots() in apfs_readdir()
Replace the calls to dir_emit_dot() and dir_emit_dotdot() with a single
call to dir_emit_dots().

These functions simply emit the vfs inode number of the current and
parent directories, so they might produce a wrong value on 32-bit
architectures.  I don't believe this can be fixed from inside the
module, because one of our directories could be '..' to another file
system.  BTRFS simply ignores the issue; do the same for now, but add a
comment on it.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>