apfs: reset dir child count on dentry cleanup
While apfs_create_dentry() itself resets the directory child count on
failure, the callers have been ignoring the issue.  This isn't very
important yet because these calls come right before transaction commit,
and failure at that stage sets the filesystem read-only.  This is not
likely to be the case in the future implementation of ->rename() though,
so define and use a new apfs_undo_create_dentry().

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>