apfs: never reuse query structures during search
For non-multiple queries, apfs_btree_query() does not really need to
remember the parent query, so I originally decided to reuse it to save
one allocation.  This leads to some ugly code, and extra complications
for inserting new records.  There is no verified performance benefit,
and even if there was, allocating from a cache could be a better idea.

Just use the approach for multiple queries on all cases.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>