apfs: call apfs_read_node() with id instead of bno
Buffer heads for b-tree nodes are typically read twice: first by the
object helpers, which may need to translate an id into a block number,
and later by apfs_read_node() itself, which is called with the obtained
block number.  This is redundant and ugly.

Redefine apfs_read_node() to be called with an object id (and a storage
type) instead of a block number; also add a 'write' argument to request
write access.  Call the object helpers inside.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>