apfs: split catalog nodes when full
Add support for catalog trees with multiple nodes in readwrite mode.
This requires splitting nodes that become full, deleting nodes that
become empty, and occasionally creating a new root with greater height.

Many corner cases exist that are difficult to test with the current
limitations of the module.  It's likely that several bugs will surface
in the future, but I prefer to press forward and get xfstests to run as
soon as possible.

Keeping track of the parent query when a node is split would add a lot
of complexity, so don't bother with that anymore.  The only reason this
mattered here was so that we could use a call to apfs_btree_remove()
followed by apfs_btree_insert() to replace a record.  Define a new
apfs_btree_replace() instead.

This patch has introduced several recursive function calls; some of them
would be very hard to get rid of, but I will look into it at some later
point.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>