apfs: keep a node's free space length in memory
When a record is removed or resized, the free space in the node becomes
fragmented.  The total length of the fragmented space in each of the key
and value areas must be reported by the node header, which is currently
updated directly by apfs_btree_remove().

To make this simpler, keep the lengths on the in-memory node structure;
make all needed modifications there, and write them to the on-disk
structure on apfs_update_node().

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>