apfs: don't take reference to inode on ->rename()
Unmounting an apfs container after renaming a file throws a variation of
the following message:

  VFS: Busy inodes after unmount of loop0. Self-destruct in 5 seconds.  Have a nice day...

The reason is that ->rename() is not supposed to increase the reference
count of the inode, but it's doing so through the call to __apfs_link().
One might expect __apfs_unlink() to revert this, but it doesn't; this is
because ->link() is required to get the inode, but ->unlink() is not
allowed to put it.

So move the call to ihold() from __apfs_link() to apfs_link() itself,
and adjust the cleanup accordingly.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>