apfs: implement creation of free queue nodes
Extend apfs_create_node() to support the creation of free queue nodes.
Define and use apfs_cpoint_data_allocate() to allocate the blocks in the
checkpoint data area, and apfs_create_cpoint_map() to register them in
the checkpoint-mapping block.  In the future I would like to simplify
apfs_create_node() with functions such as "apfs_create_ephemeral_object"
and "apfs_create_virtual_object" that took care of this stuff.

The module will now be able to perform far more writes before getting
an error, which will simplify testing.  We still need to put a limit to
the size of the free queue tree, and actually free the blocks in the
bitmap.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>