apfs: don't write overridden uids/gids to disk
If the uid or gid mount options are in use, any existing file that gets
touched will lose its original ownership data, which is undesirable.

Follow JFS in this matter, as we've always done so far.  When an inode
is read or created, store its uid/gid in the new i_saved_uid/i_saved_gid
fields.  If either of the mount options is in use, apfs_update_inode()
will use these apfs_inode_info fields instead of the vfs inode's.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>