apfs: define functions for xfield creation
Renaming an inode creates a new record from scratch, so all existing
extended fields are lost.  This hasn't been a problem so far for files
created by the module, because the only two supported xfields are always
rebuilt.

To do things correctly, move the creation of extended fields to the new
apfs_insert_xfield(), which is capable of handling xfield replacement as
well.

Also define apfs_init_xfields() to initialize the blob for an xfield
collection; use both functions to simplify apfs_build_dentry_val() and
apfs_build_inode_val().

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>