apfs: rename macro for #include guard in extents.h
By mistake, the macro guard of extents.h was given a name that is prone
to collision.  Rename it to specify the module.

This change should probably have happened as part of commit 5f7efae269f5
("apfs: rename macro for #include guard in object.h"), but I seem to
have missed it back then.

Fixes: 7132a98d7fce ("apfs: move extent code to separate files")
Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>