apfs: fix endianness issue for allocation bitmaps
Before working with an on-disk allocation bitmap, apfs_chunk_find_free()
implicitly casts it to a u64 array, ignoring the cpu endianness.  Since
the kernel provides find_next_zero_bit_le(), get rid of my buggy code
and use that instead.

To prevent this in the future, don't declare the bitmap as void* in the
caller.

Fixes: 3bdb0a803208 ("apfs: implement block allocation")
Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>