apfs: use __set_bit_le() to set bitmap bits
In the spirit of the previous patch, use the __set_bit_le() provided by
the kernel to simplify apfs_chunk_mark_used().  There is no endianness
issue this time, but it's still better to use the common code.

Also change the type of the bitmap from u8* to char*, for consistency
with apfs_chunk_find_free().

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>