apfs: don't do CoW for new blank chunk bitmaps
When apfs_chunk_allocate_block() is called on a chunk that is entirely
free, a blank bitmap must be allocated.  This new bitmap shouldn't go
through copy-on-write, but it does; the result is that it gets put in
the free queue immediately, corrupting the container.

No consequences have been noticed so far, because this module can't yet
allocate enough blocks to fill a whole chunk.  Fix the issue anyway; the
patch can still be tested if we tweak the code so that it never performs
any allocations in the first chunk.

Fixes: 3bdb0a803208 ("apfs: implement block allocation")
Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>