apfs: fail cpm block creation with EOPNOTSUPP
Creation of cpm blocks is not yet supported, which may cause free queue
insertions to fail.  Currently the return code used is ENOSPC, but this
has a special meaning for the spaceman allocation functions: it forces
them to continue the search in the next chunk or cib, with unpredictable
consequences that seem to trigger a few assertions.  To avoid this, use
EOPNOTSUPP instead.

Fixes: 06af0392cf42 ("apfs: implement creation of free queue nodes")
Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>